requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
antlr4-python3-runtime=4.9.3=pypi_0
ants=2.3.1=hf484d3e_0
antspyx=0.4.2=pypi_0
anyio=4.4.0=pypi_0
appdirs=1.4.4=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
async-lru=2.0.4=pypi_0
attrs=23.2.0=pypi_0
babel=2.15.0=pypi_0
backcall=0.2.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=mkl
bleach=6.1.0=pypi_0
brotli-python=1.0.9=py38h6a678d5_7
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.12.12=h06a4308_0
certifi=2023.11.17=py38h06a4308_0
cffi=1.16.0=py38h5eee18b_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
chart-studio=1.1.0=pypi_0
click=8.1.7=pypi_0
colorama=0.4.6=pypi_0
comm=0.2.2=pypi_0
contourpy=1.1.1=pypi_0
cryptography=41.0.7=py38hdda0065_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cycler=0.12.1=pypi_0
cython=3.0.6=pypi_0
debugpy=1.8.1=pypi_0
decorator=5.1.1=pypi_0
deepdiff=6.7.1=pypi_0
defusedxml=0.7.1=pypi_0
deprecated=1.2.14=pypi_0
dill=0.3.7=pypi_0
dipy=1.8.0=pypi_0
docker-pycreds=0.4.0=pypi_0
et-xmlfile=1.1.0=pypi_0
exceptiongroup=1.2.1=pypi_0
executing=2.0.1=pypi_0
fastjsonschema=2.19.1=pypi_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.13.1=py38h06a4308_0
fonttools=4.46.0=pypi_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=h4a9f257_0
fslpy=3.15.4=pypi_0
fsspec=2023.12.2=pypi_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.11=pypi_0
gitpython=3.1.40=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
h11=0.14.0=pypi_0
h5py=3.10.0=pypi_0
httpcore=1.0.5=pypi_0
httpx=0.27.0=pypi_0
humanize=4.9.0=pypi_0
hydra-core=1.3.2=pypi_0
idna=3.4=py38h06a4308_0
imageio=2.33.1=pypi_0
importlib-metadata=7.1.0=pypi_0
importlib-resources=6.1.1=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
ipykernel=6.29.4=pypi_0
ipython=8.12.3=pypi_0
ipywidgets=8.1.3=pypi_0
isoduration=20.11.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.2=py38h06a4308_0
joblib=1.3.2=pypi_0
jpeg=9e=h5eee18b_1
json5=0.9.25=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.6.2=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.1=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.2.1=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.2=pypi_0
jupyterlab-widgets=3.0.11=pypi_0
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lazy-loader=0.3=pypi_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.8.1.2=0
libcurand=10.3.4.101=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_0
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libopus=1.3.1=h7b6447c_0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libvpx=1.7.0=h439df22_0
libwebp=1.3.2=h11a3e52_0
libwebp-base=1.3.2=h5eee18b_0
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markdown-it-py=3.0.0=pypi_0
markupsafe=2.1.1=py38h7f8727e_0
matplotlib=3.7.4=pypi_0
matplotlib-inline=0.1.7=pypi_0
mdurl=0.1.2=pypi_0
mistune=3.0.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.8=py38h5eee18b_0
mkl_random=1.2.4=py38hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py38h06a4308_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.6.0=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py38h06a4308_0
nibabel=5.1.0=pypi_0
notebook=7.2.0=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
omegaconf=2.3.0=pypi_0
opencv-python=4.8.1.78=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openpyxl=3.1.5=pypi_0
openssl=3.0.12=h7f8727e_0
ordered-set=4.1.0=pypi_0
overrides=7.7.0=pypi_0
packaging=23.2=pypi_0
pandas=2.0.3=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
patsy=0.5.4=pypi_0
pexpect=4.9.0=pypi_0
pickleshare=0.7.5=pypi_0
pillow=10.0.1=py38ha6cbd5a_0
pip=23.3.1=py38h06a4308_0
pkgutil-resolve-name=1.3.10=pypi_0
platformdirs=4.2.2=pypi_0
plotly=5.18.0=pypi_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.45=pypi_0
protobuf=4.25.1=pypi_0
psutil=5.9.6=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.17.2=pypi_0
pyopenssl=23.2.0=py38h06a4308_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=py38h06a4308_0
python=3.8.18=h955ad1f_0
python-dateutil=2.8.2=pypi_0
python-json-logger=2.0.7=pypi_0
pytorch=2.1.1=py3.8_cuda11.8_cudnn8.7.0_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytz=2023.3.post1=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=py38h5eee18b_0
pyzmq=26.0.3=pypi_0
qtconsole=5.5.2=pypi_0
qtpy=2.4.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
requests=2.31.0=py38h06a4308_0
retrying=1.3.4=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
rich=13.7.0=pypi_0
rpds-py=0.18.1=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.3.2=pypi_0
scipy=1.10.1=pypi_0
send2trash=1.8.3=pypi_0
sentry-sdk=1.38.0=pypi_0
setproctitle=1.3.3=pypi_0
setuptools=68.2.2=py38h06a4308_0
setuptools-scm=8.0.4=pypi_0
shellingham=1.5.4=pypi_0
simpleitk=2.3.1=pypi_0
six=1.16.0=pypi_0
smmap=5.0.1=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.41.2=h5eee18b_0
stack-data=0.6.3=pypi_0
statsmodels=0.14.0=pypi_0
sympy=1.12=py38h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tenacity=8.2.3=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.2.0=pypi_0
tifffile=2023.7.10=pypi_0
tinycss2=1.3.0=pypi_0
tk=8.6.12=h1ccaba5_0
tomli=2.0.1=pypi_0
torchaudio=2.1.1=py38_cu118
torchio=0.19.3=pypi_0
torchtriton=2.1.0=py38
torchvision=0.16.1=py38_cu118
tornado=6.4=pypi_0
tqdm=4.66.1=pypi_0
traitlets=5.14.3=pypi_0
trx-python=0.2.9=pypi_0
typer=0.9.0=pypi_0
types-python-dateutil=2.9.0.20240316=pypi_0
typing_extensions=4.7.1=py38h06a4308_0
tzdata=2023.3=pypi_0
uri-template=1.3.0=pypi_0
urllib3=1.26.18=py38h06a4308_0
wandb=0.16.1=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
wheel=0.41.2=py38h06a4308_0
widgetsnbextension=4.0.11=pypi_0
wrapt=1.16.0=pypi_0
x264=1!157.20191217=h7b6447c_0
xz=5.4.5=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zipp=3.17.0=pypi_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.5=hc292b87_0
